schedule: split balance_test.go
