+++ pkg/schedule/schedulers/scatter_range_test.go
@@ -1,0 +1,99 @@
+// Copyright 2024 TiKV Project Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package schedulers
+
+import (
+	"fmt"
+	"math/rand"
+	"testing"
+
+	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/stretchr/testify/require"
+	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/schedule/types"
+	"github.com/tikv/pd/pkg/storage"
+	"github.com/tikv/pd/pkg/versioninfo"
+)
+
+func TestScatterRangeBalance(t *testing.T) {
+	re := require.New(t)
+	checkScatterRangeBalance(re, false /* disable placement rules */)
+	checkScatterRangeBalance(re, true /* enable placement rules */)
+}
+
+func checkScatterRangeBalance(re *require.Assertions, enablePlacementRules bool) {
+	cancel, _, tc, oc := prepareSchedulersTest()
+	defer cancel()
+	tc.SetClusterVersion(versioninfo.MinSupportedVersion(versioninfo.Version4_0))
+	tc.SetEnablePlacementRules(enablePlacementRules)
+	tc.SetMaxReplicasWithLabel(enablePlacementRules, 3)
+	// range cluster use a special tolerant ratio, cluster opt take no impact
+	tc.SetTolerantSizeRatio(10000)
+	// Add stores 1,2,3,4,5.
+	tc.AddRegionStore(1, 0)
+	tc.AddRegionStore(2, 0)
+	tc.AddRegionStore(3, 0)
+	tc.AddRegionStore(4, 0)
+	tc.AddRegionStore(5, 0)
+	var (
+		id      uint64
+		regions []*metapb.Region
+	)
+	for i := 0; i < 50; i++ {
+		peers := []*metapb.Peer{
+			{Id: id + 1, StoreId: 1},
+			{Id: id + 2, StoreId: 2},
+			{Id: id + 3, StoreId: 3},
+		}
+		regions = append(regions, &metapb.Region{
+			Id:       id + 4,
+			Peers:    peers,
+			StartKey: []byte(fmt.Sprintf("s_%02d", i)),
+			EndKey:   []byte(fmt.Sprintf("s_%02d", i+1)),
+		})
+		id += 4
+	}
+	// empty region case
+	regions[49].EndKey = []byte("")
+	for _, meta := range regions {
+		leader := rand.Intn(4) % 3
+		regionInfo := core.NewRegionInfo(
+			meta,
+			meta.Peers[leader],
+			core.SetApproximateKeys(1),
+			core.SetApproximateSize(1),
+		)
+		origin, overlaps, rangeChanged := tc.SetRegion(regionInfo)
+		tc.UpdateSubTree(regionInfo, origin, overlaps, rangeChanged)
+	}
+	for i := 0; i < 100; i++ {
+		_, err := tc.AllocPeer(1)
+		re.NoError(err)
+	}
+	for i := 1; i <= 5; i++ {
+		tc.UpdateStoreStatus(uint64(i))
+	}
+
+	hb, err := CreateScheduler(types.ScatterRangeScheduler, oc, storage.NewStorageWithMemoryBackend(), ConfigSliceDecoder(types.ScatterRangeScheduler, []string{"s_00", "s_50", "t"}))
+	re.NoError(err)
+
+	scheduleAndApplyOperator(tc, hb, 100)
+	for i := 1; i <= 5; i++ {
+		leaderCount := tc.GetStoreLeaderCount(uint64(i))
+		re.LessOrEqual(leaderCount, 12)
+		regionCount = tc.GetStoreRegionCount(uint64(i))
+		re.LessOrEqual(regionCount, 32)
+	}
+}
